Floss: force connect to both A2DP and HFP if available
Some headsets do not connect to all available profiles in certain common
scenarios (e.g., initiating connection via power button in XM4 only
connects HFP).

In this CL, we enforce the policy to force connect to the other profile
as soon as one is connected, and expect it to be completed in 5s.

Bug: 240641968
Tag: #floss
Test: Build and verify all profiles are good for various headsets
Change-Id: Id4e2dd5ebb4d53cd7ccac9a4c12f6c4646da09f7