Floss: Implement set allowed services in BluetoothAdmin
This API disables the services that is not in the allowed list. A
profile is disabled by calling |disable| in topshim layer which should
call |cleanup| to the libbluetooth interface. Similarly, a profile is
re-enabled by calling |enable| in topshim and |init| of the libbluetooth
interface will be called subsequently.

However, the cleanup function of HFP and A2DP in topshim-cpp layer
haven't been implemented. By testing locally, the consequence is that
when a2dp and hfp are disabled by policy, the audio headset can still be
connected and stay connected, but there is no audio stream to the
headset nor the chromebook. Additionally, the headset could adjust the
volumn via HFP.

Bug: 239470589
Test: test it along with the top of this chain
Tag: #floss
Change-Id: Id4eba16f298599dad8d469bf719bc5ef2211c639