bluetooth: Fix checking Le Audio support.
When checking for LeAudio support, we should not check if BLE is
currently ON as this is inrelevant.
Some of the services, like Telecom, getProfileProxy on init, and waits
for service connection. With current code, Telecom will never get
LeAudioService instance, and will not allow to use LeAudio devices for
the phonecalls.

Bug: 243847810
Test: atest BluetoothInstrumentationTests
Tag: #feature
Change-Id: I9d82b1c7b87edbe7c1323039f1fe65e6d286af4a