Floss: update eSCO params for offloading
Per spec in 7.1.45 Enhanced Setup Synchronous Connection command | BT Core v5.3, there is an excerpt: "The Input_Transport_Unit_Size and Output_Transport_Unit_Size indicate how many bits are in each unit of data delivered by the audio data transport".

While it suggests this number usually be 8 or 16, it is meant to be vendor-specific and in previous WIP patches we were advised by Intel to go with 1.

Bug: 237373343
Tag: #floss
Test: emerge-brya floss
Change-Id: I6226434c842ec3c01ecd8eaa2faa84a0b00363c7