Move enabling state of Bluetooth sharing to system service.
BluetoothOppLauncherActivity & BluetoothOppBtEnableActivity
'enabled' state should be updated wether Bluetooth is ON or
OFF. Previously, the state of these Activities was set with
resource config, but since Bluetooth started replacing them
by sysprops, we had to enable / disable these Activities in
BluetoothOppService. However, the bt sharing tile should be
available even when Bluetooth is OFF, and as their state is
already updated in BluetoothManagerService when there is an
user change, we can use that instead but with the condition
that BluetoothOppBtEnableActivity is enabled too. All other
Bluetooth OPP Activities enabled / disabled state are still
handled in BluetoothOppService as they require Bluetooth to
be ON to work anyway, and don't impact UI tiles. This fixes
both an issue where the bt sharing tile doesn't appear when
Bluetooth has been turned OFF, and an issue where Bluetooth
would crash if it hadn't been turned ON before using the bt
sharing tile.

Test: manual
Tag: #feature
Bug: 242963176
Merged-In: I67840dbd9f4cfc42026a14a1fc467fc95aa3a587
Change-Id: I67840dbd9f4cfc42026a14a1fc467fc95aa3a587