[RootCanal] Add heartbeats to handle remote disconnections
Ping the peer if we haven't heard anything from it in a while. Respond
to all pings with an ignored PING_RESPONSE. If no packet is received for
some duration, drop the connection.

The LSTO duration is not currently configurable. We will also "panic"
and keep sending pings if the peer takes a while to respond, but that's
probably fine (since ticks aren't that frequent).

Bug: 243698555
Tag: #feature
Test: atest pts-bot on next CL
Change-Id: Id5f7310c5322adfd8548953228ab44c6e7d861b8