RootCanal: Fix LE advertising with legacy HCI commands
When using the legacay advertising commands (a.k.a. non extended command set),
the address used for advertising messages may not be set, resulting in
scanning and connection failures for the peer device.

This commit only addresses this particular issue, larger changes are
required to solve other latent problems: http://b/228326164

Test: GATT PTS tests against Eiffel
Bug: 228326164
Change-Id: Ib5b60683abbbe1063387bff6a59a58e981a914d3