Revert "Bass: Fix BIS sync state channel map"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9c3c5d40-7df2-4a93-9eff-410c1c90db2b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e06c635a8b470356eda565cce81d9296848e82d