Revert "LeAudioTestApp/Bass: Allow channel selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9c3c5d40-7df2-4a93-9eff-410c1c90db2b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d820291d64b7e419dbb91b968db85deb0c3c2eb