floss: Change default adapter based on system state
The default adapter on the system is the lowest numbered, present
adapter. This will usually be hci0 which is the first Bluetooth adapter
that starts up with the system. However, that adapter can sometimes
rename itself as hci1 if the adapter is lost and probed very quickly
(before userspace can close the socket).

This change gets around this by setting the default adapter to always be
either the a) desired hci adapter or b) the lowest numbered available
adapter.

This is still not a good solution for multiple adapters (which should
actually use full hardware path rather than hci index) but is useable
in the interim.

Bug: 244457509
Tag: #floss
Test: Disable internal BT and plug in external BT

Change-Id: Id4c01eafc23b0dc8f57c6ff970b98473c81c5ca2