Fix rc connect fail after remove bond
[Description]
when DUT is direct connect to RC, user remove bond of the RC.
whatever the RC connected or disconnected, it will not notify
framework the connection status, because bta_hh device will
be clear when remove bond.
so when remove bond and pending device is the connecting device,
then notify the framework disconnected, and cancel direct connect.

Bug: TBD
Test: connect successfully after remove bond.

Change-Id: Ib1a23c5bcd076e2e0a395b6b9a10fca90355d0de