Revert "leaudio: Synchronize TBS and LeAudio"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3495d82b-82ab-49db-9142-d3137305ff94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa4b929fdd9f95c1ba8f465c2adb46c54d7e483b