Seperate audio allocation and connection status configuraiton update
With the patch, we change flow as the downmix_fallback_ set to true via
persist.bluetooth.leaudio.offloader.downmix_fallback. As the stream
started, the bluetooth stack will update the real left/right audio
allocation to the bluetooth audio HAL. If part of cises are not
connected, the stack will update the audio configuration again with the
streamMap that contains the connection state.

Bug: 231084798
Test: check the audio configuration update sequence as downmix_fallback_
enable/disable
Test: atest packages/modules/Bluetooth

Change-Id: Ic5d4bb8b3f265499657f2d162b3616e1b377adbb