Separate audio allocation and connection status configuration update
With the patch, we change flow and remove downmix property. As the stream started, the bluetooth stack will update the target
allocation to the bluetooth audio HAL. If part of cises are not
connected, the stack will update the audio configuration again with the
streamMap that contains the current allocation.

Bug: 231084798
Test: check the audio configuration update sequence as downmix_fallback_
enable/disable
Test: atest bluetooth_le_audio_test

Change-Id: Ic5d4bb8b3f265499657f2d162b3616e1b377adbb