floss: Recover via reset after multiple starts
If we attempt to start multiple times and are unsuccessful, try using
the reset mechanism to recover instead of trying to start.

This also changes the condition for attempting a restart/reset from
depending on device presence. On Linux, when we bind the socket as
userchannel, the device presence is false (index is removed). It makes
more sense to attempt retry if we want the device enabled.

Bug: 236709385
Tag: #floss
Test: Enable/disable cycles to bad state and trigger reset.
Change-Id: I8a9003385f8fcb6d1ac237670087d34ed9590048