Revert "[test] fix GoogleBluetoothInstrumentationTests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b1f6c68e-6f26-40a8-87bf-8dc8e9ec749b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie228070099f82dd17ac926d18833d85bb1d0f283