Revert "[Pandora] Move pairing interface into a separate interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/1f2d4f12-35e1-4fc7-b828-db6891b0531d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f8d29bc07eda5e03018549100b8f3bef3f258f5