btif: hh: emit (DIS)CONNECTING state when a remote device (dis)connect
When a remote HID device initiates the connection to the host, a
BTA_HH_OPEN_EVT or a BTA_HH_CLOSE_EVT will be triggered, and the upper
layer will receive a CONNECTED or DISCONNECTED state callback upon
finishing handling this event. Unlike the connections initiated by the
host, where the attempt is identifiable, the remote devices triggered
only have status updates when connected or disconnected.

This patch has the hid host to upstream (dis)connecting event whenever
received the attempt from devices to help with the logging.

BUG: 240781725
Tag: #floss
Test: gd/cert/run

Change-Id: Iff5e851fb3c98d991cd81d209af140e92937ce7d