Floss: use CLOCK_MONOTONIC_RAW in A2DP encoding for non-Android builds
Floss uses CLOCK_MONOTONIC and CLOCK_BOOTTIME for various timestamp and
timediff calculation.

This causes glitch when the audio stack calculates timestamp and
schedules with CLOCK_MONOTONIC_RAW.

In this CL, we swap out the clock used in scheduling A2DP packet
encoding to use CLOCK_MONOTONIC_RAW when the build is for non-Android.

Bug: 241816575
Tag: #floss
Test: Build and verify A2DP doesn't glitch

Change-Id: I275208e33285282911c49caa599cbb14a810bfa9