Revert "Cache EIR UUIDs per Bluetooth address outside the pairin..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/ab9951e7-8958-469e-a6dd-892fdca16958). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31d79beec2460bf0feddd89a8deeb2458d2b878a