Revert "Ignore empty as well as base UUIDs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/ab9951e7-8958-469e-a6dd-892fdca16958). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc1b07d4a955fc85f6c316c74b7ee591590b6942