floss: removes the API UnregisterAdvertiserCallback
This API is not required as Floss daemon will do the same thing while
the D-Bus clients leave. As it is not being used by chromium and
btclient, we would like to remove this redundant API and avoid it's
being used in unexpected ways, such as stopping all advertising sets.

Bug: 237346241
Bug: 246474368
Tag: #floss
Test: ./buid.py --target test
Test: btclient advertise [on|off]
Change-Id: I7be6835c93e006453876d3700e8262ed1bbd2129