LeAudio: Refactor le audio hal client
This splits the client_audio.cc for sink and source files
and moves them into a new subdirectory for better separation
and mantainability. It was also put into the le_audio namespace.

It simplifies class naming - now named from the perspective of
the bluetooth profile role rather than mixing the bluetooth
domain with the framework point of view. This eliminates the
non-intuitive usages like registering LeAudioClientAudioSinkReceiver
in LeAudioClientAudioSource. It could be confusing who is Sink,
who is Source and what the Receiver is. After this change, the
AudioHalClient::LeAudioSourceCallbacks instance is registered
in AudioHalClient::LeAudioSource.

The interface was cleaned up, simplified and the confusing
const void* token has been removed. The object ownership was
redefined and the AudioHalClient singleton manages the
Audio Hal Sink/Source endpoints acquiring, releasing and
object allocation and deallocation. Previously this was done
in multiple places, which causes situations like le audio service
allocating and releasing it in the production run, but only releasing
it in the unit test runs. But in some test cases this object was
deallocated by the test case itself. This required special handling
to avoid double free or using stack allocated objects.

Lastly, it improves logs and switches to the new logging mechanism.

Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster --no-bazel-mode
Bug: 249737696
Tag: #refactor
Change-Id: I41da290d7941b4a1a3001015cb9c15a89f9c3375