Floss: Fix usage of BT_STATUS_BUSY
Since we want to expose the `btif_status_t` as public API for Floss, we
need to be consistent in the interpretation of the error codes.

Changed error code in btif_hf and btif_hd to use BT_STATUS_DONE instead
of BT_STATUS_BUSY since those cases look like the errors are not
retryable, but rather a request that is already completed and can be
assumed success unless need to be redone.

Bug: 247110650
Tag: #floss
Test: Manual - Build Floss on Linux

Change-Id: If483dff010a620cafe9f659cfdc1ac8d7af37025