Fix ACL Create Connection Cancel
The old cancellation implementation has two issues: (1) if a connection
request is queued, we still tell the controller to cancel it and then
dequeue the request later, which makes no sense, and (2) we *don't*
cancel the request if it is outgoing, because the predicate is flipped
from what it should be.

This fixes both and integrates it with the ACL scheduler.

Bug: 246640776
Test: TODO
Tag: #stability
Change-Id: Ibf82f076497738a0d885a0097862ff926d5e4605