leaudio: Workaround for the CIS->SCO handover
Some controllers have an issue when doing the handover from CIS to SCO.
For this purpose let's sync a bit and notify upper layer when LeAudio
group turned idle if it is on call.

This functionallity is hidden behind
persist.bluetooth.leaudio.notify.idle.during.call

for the moment. To enable notification, the above shall be set to true.

Bug: 242908683
Test: atest BluetoothInstrumentationTests
Tag: #feature
Change-Id: I2bdd81612f0dd06d77ed44a08c17545121bf6c33