Revert "leaudio: Improve inactivating LeAudio device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9cb18bea-6ea3-4e01-8a9c-b37cf6e9d77c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90afd32bd96e4d6a6aa4246f572b2284db093711