floss: Send ACL connection state metrics
This is called upon ACL connection and disconnection, recording the
status, direction, and initiator of the transaction.

One caveat in Floss is that we don't know the transport to be used
(classic | le) on initiating a connection, therefore we don't
immediately report the "connecting" event, but rather group it
together with the "connected" event, where the transport is conveyed
to floss.

For symmetry simplicity, that is done as well for disconnection.

Bug: 240782154
Tag: #floss
Test: Verify the ACL send metric message is correctly logged

Change-Id: I564b0f7be2addf7ce33be2efe8946eecc33fb3de