leaudio: Recover on CIG not being removed on BT OFF
If BT OFF is done while ISO streming, stack might not have time to
remove CIG, as there are still CISes avalable.
Current stack design does not allow to wait for CIS to be disconnected
in such a case, therefor this patch, which tries to recover when
Creating CIG will endup with COMMAND DISALLOWED.

Note: This happens only when BT Controller is not Reseted on BT OFF due
to other activities.

Bug: 228351333
Bug: 246304525
Tag: #feature
Test: BluetoothInstrumentationTests
Change-Id: I0a86e8851491cf147fe6f84ddd86be3466ef6c64