Revert "Delete keys if SMP over BR fails"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/219a7fd4-571e-4e55-bfc8-446310f0ff2f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6694db716f0c2e8179913c521b8f5291361a6e75