Revert "pdl: Use correct semantic for padding field"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/8a92b115-aa73-4fe6-9987-8b00e8bd5034). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic418e7ef45d04271949e8643abcb1d1df5b4322b