Revert "leaudio: Notify connected when subscribed for CCC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/51ffc743-099c-4a46-8477-f4dfb14cc7a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iafa5a0430658c40456ace56a4c1c41b6fb7d027e