Revert "leaudio: Notify connected when subscribed for CCC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/383a5a1d-d50f-400a-8b32-2eecf827987e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifaccea8612d3250e5c845daf8023106d8a737a34