VolumeControl: Do not set the same volume level twice
Devices may connect or be discovered as a new group
memeber after the streaming was already started, or
services can be connected in a different order each
time, the newly connected device may need additional
volume level corrections. Some of them may already
be pending on the native layer due to previous group
API calls, while Java has not yet been notified about
all the native stack events.

This change eliminates any additional volume state
changes if the volume is already set to the desired
level but the Java layer does not know that yet, or
when the volume changing GATT operation with same
parameters is already on the operation queue.

Bug: 248969553
Bug: 248915809
Tag: #feature
Test: atest --host bluetooth_vc_test --no-bazel-mode
Change-Id: Ic7ad61a8f8be6a48bb09aa3b07059e8f1abbd528