VolumeControl: Set proper volume on joining device
When one of the devices connects to an already streaming group it
needs to have it volume level corrected to the already cached group
volume level.

Additionally on the initial connection when the second device is not yet
recognised by CSIS as a group member but the Volume Control Service
has already been connected and group volume was already set by the
LeAudioService, the newly added group member also needs his volume level
to be corrected to the already set group volume level.

Bug: 248969553
Bug: 248915809
Tag: #feature
Test: atest BluetoothInstrumentationTests
Change-Id: I02e1827c4f652f78ab86365b7b9e72ef0c1c77c2