Revert "leaudio: Do not start stream if there is pending reconfi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b4ac9e09-14cf-429c-bedb-45ee7785f662). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39e3b13f715f15aa2d4091bf3649174e848d2c93