Replace CTKD detection with information from lower layers
When performing CTKD, we know about it in SMP, but don't pass this
information to BTA and BTIF layer. In BTIF, we guessed wether CTKD
happens by checking if address returned from pairing is different than
one we requested pairing on. This assumes device is using RPA or Static
address over LE. This is incorrect - device can advertise using Public
address over LE, and still perform CTKD.

Because of this incorrect assumption, we would perform SDP service
discovery when we just generated link key, rather than waiting for end
of bonding. This in turn caused a race condition with
delayed_auth_timer_ent - timer that waits 0.5s before pairing is
considered done.

Bug: tbd
Test: pair with device using Public transport over LE multiple times
     ensure device is paired and visible in settings.
Change-Id: Iacefcd0439d7fba1c1b1ab873f79c6aae9c73eb8