Revert "le_audio: Disable ASEs after 3sec from suspend request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/03b0715b-b1a6-4e05-8074-e2e61143f521). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic648b6edc88a4af0e0b75eb41210618b1f4fef57