Revert "le_audio: Disable ASEs after 3sec from suspend request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d6dc09c2-556c-42cd-a4a8-f5d4366f6bce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic49b17c6e0e16501f165dade9c1663d47dd54b9b