Revert "le_audio: Disable ASEs after 3sec from suspend request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/da12a3a7-c84b-4a1f-8a4e-9fc2a87b449a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I746776f732d5c3786f226d51d68a4626f0c6c454