Revert "le_audio: Disable ASEs after 3sec from suspend request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/8ad8ee39-ff3e-4bc3-b017-da6d29fc05f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77aace8cfafea977794462df3161af07de464ce1