Revert "le_audio: Disable ASEs after 3sec from suspend request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4a5c3b8e-c729-4860-9259-3826ddefd838). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie238b2a14f897e614d232ff60d3c07539809410c