[PTS-Bot]:
Draft implementation to support accessing Browsing folder, Notification registration and Media player selection commands

Added file AvrdpBrowserService
This is a service which extends "MediaBrowserService" with MediaSession and initial implementation to support browsing folders.
With registration to Media buttons callback which responds to PTT commands(Tried with Play/Pause..).
Currently the folder structure is Root folder with 5 meta data items.

TODO
To generalize setting the playback state and metadata according the MMI request.
Creating folder hierarchy for browsing test cases.

Request to confirm if we can proceed with this approach.

Test: atest pts-bot:AVRCP -v
Change-Id: I19085b8613c74417b614793a4958927c638cf92a