Floss: Add Uuid::from([u8; 16])
It's natural to build Uuid from an array of 16 bytes. Therefore it's
good to have a standard conversion utility that does not involve vector
overhead and also callers don't need to know the internal field backing
the array.

Also includes a trivial warning fix for unnecessary mut variable.

Bug: 176837458
Tag: #floss
Test: Build Floss on Chrome OS and Linux

Change-Id: Idffeec4f55f69a6bd0769c52a5ea2a5243987417