pdl: Add support for ‘enum’ declarations
This only allows us to parse the enums and to generate Rust code for
the enum declaration itself. Support for enums inside packets will
come next.

Tests have been added to the canonical PDL files: they demonstrate
that we can generate valid Rust code. A test was added to show that we
need “#[repr(u64)]” for an enum with a maximum discriminant.

Test: atest pdl_rust_generator_tests_{le,be} pdl_tests
Change-Id: I4f8777c17bc0d7ddeddf214c66db03f070d3794e