Make pause/stop cancel playback reestablishment during transient loss
Our goal in a transient loss is to pause the remote player as a courtesy
and then attempt to reestablish playback once we gain focus back.
However, in rare cases, an application (like assistant) can request and
send a pause while we're in a transient loss and already paused. This
should cause playback to stop, however we don't currently react to the
pause and cancel our note to reestablish playback.

This change moves the play/pause decision logic to AVRCP Controller, and
allows local requests to transport controls to cancel our future attempt
to establish playback out of a transient loss if a user sents a pause or
stop.

Tag: #stability
Bug: 254536123
Test: atest BluetoothInstrumentationTests
Change-Id: I277e35f7b432fc7d287cbd4ab19c2ef5f3b870bb