Revert "RootCanal: Configure the default LE vendor capabilities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9f0ef993-3be9-413f-bb01-e57f8b5bd11f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0812a18fb75c648e27fa615e8bd959bd8b2be032