Use test app's data folder as filter directory
During a retry, test methods to be retried are written into a
file, pushed onto device and read by the test instrumentation
for execution.

The default location for such filter file is /data/local/tmp,
which is prohibited for access by system uid packages, such
as the system UI tests package.

This change uses the test app's data folder for filter: this
is not generally applicable for all the tests since root
permission is required, but it is a relatively safe assumption
in case of testing against system uid pakcages since it
implies userdebug build is being used.

Bug: 255088822
Test: atest BluetoothInstrumentationTests
Change-Id: I8e1239d9715e127108ce67365ea8ee4af6663a1b