Revert "Add BluetoothOppUtilityTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/59ae2d63-3649-4085-ae26-fcbe3d3c9549). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb626e94aca4cdf85b4c9f3df662fb166ef6cfe0