Fix infinite L2CAP timeout when ATT is used without GATT clients
If we do ATT operations without a GATT client, the channel timeout stays
at the default value (infinite) which prevents the ACL link from
disconnecting.

The timeout is set to zero when the *last* GATT client disconnects, but
of course this doesn't work if there never are *any* GATT clients, yet
ATT operations (such as service discovery) are taking place.

Test: atest pts-bot?
Bug: 251954527
Change-Id: I2990fd8892ab0f442fd7847b34812e29d832baac