LeAudio: Refactor context type management functions
While the previous refactor was focused on the types
used for passing audio contexts, this one takes the
code logic into the account and improves the readability
and the concepts used in the code.

It renames functions around the context type management.
The 'active context' term was ambigious and was replaced
with 'available context', which the BT spec. is using.
Functions now use more coherent naming scheme and since
some of them were returning a bit masks which were
not bit-tested but logically ANDed with other booleans,
they now return boolean type to avoid the confusion and
implicit conversion bugs to happen.

Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine  --no-bazel-mode
Bug: b/249737696
Tag: #refactor
Change-Id: I4305dc64c04bedb904d5e23ef29f5282a27a2860