Revert "RootCanal: Rewrite LE connection logic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d919f0c5-4ff4-41b0-b22e-26787451fb9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7df036fe41cc665cc413ed53c590afd16ab48e63