Floss: use LOG_* in a2dp_encoding_host
In Floss, there are many types of logging, one of which is the LOG(*)
format that goes through libchrome and doesn't show up in syslogs.

That is making debugging very hard as there is no way to see those logs
in feedback reports.

In this CL, we start from a2dp_encoding_host.cc to change all LOG(*)
variants to LOG_*.

Bug: 254808917
Tag: #floss
Test: Build and verify
Change-Id: I204f08f49a7b79d525c2e37563dfd1829f621d3b