Floss: replace libchrome logs with osi logs in a2dp functions
We have a mixed usage of logs, some in the form of LOG(*) and others in
the form of LOG_*, which is very confusing.

In the process of bringing-up Floss, we found that there was no way of
seeing libchrome logs and that sets a significant offset to debugging.

In this CL, we replace some of the most interesting logs that could help
debugging A2DP issues to osi logs.

Bug: 254808917
Tag: #floss
Test: Build and verify

Change-Id: I507b3c79d70b877ea477f3fcb491997e9476fec6