Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9a495115-ef00-44db-b754-de1fb7bac11a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33432a4a453141a6510da930bc201743d64c6d65