Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a948b723-adbd-4674-ae1b-ba4e44640dc0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e27a51c58439717207bca97699639f2827ccd11