Revert "Checks the service is supported before binding the service"
Revert "Checks the service is supported before binding the service"

Revert "Checks the service is supported before binding the service"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a970f06f-2682-4d54-99b3-e8775656a610). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2233c3bc1bb28b689b838b4e27e40bdead1ab697