Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/101061f3-e872-400c-9ef5-fddf4521d1f4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I079de7545f02d0b5d6f0aabcef521b58daf620a8