Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/c82b2e10-c5e5-4121-9ce4-7d19d3aeaff1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibcd4685e6e53f98ea74ad19e1450c20b4ad1f693