Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/8bb34da4-bced-45f6-a2c3-6bdba33bbf18). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22eab92f4588f21a53de5eabf71d02cf0903d268