Revert "Checks the service is supported before binding the service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b040998a-61fc-43f5-98a9-73c456a066cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaab5b930b1e7f22bd182a6d16e5b52438bb85145