[Rootcanal] Add test channel API to set feature bits on devices
This lets us better simulate controllers with different feature support.
This is more general than supporting the particular VSC used by pts-bot
as it works for Avatar too. The JSON files currently used provide a
global config, but this is not useful when we want multiple devices to
have different feature bits.

Bug: 237447510
Test: atest pts-bot
Change-Id: I1940c4d049a1c428db08b4aba45a3ace0b5301bb