Revert "BluetoothMetrics: Upload Android device HCI and LMP versions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/68b66907-a00a-4dc9-9c16-5157d380e53f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90eefc4a1ecb65a5977beb523c5bdc05afee12ec