Floss: disallow replacing callbacks in reactables
We found that some callbacks bound to member functions of |HciHalHost|
can get freed while to be run, due to |Reactor::ModifyRegistration|
allowing to replace the callback from aside without measures to prevent
the race.

Currently, |HciHalHost| is the only caller to that function, and an
approach is to disallow replacing the callbacks, and only update the
poll event masks of interest.

Bug: 256533348
Tag: #floss
Test: Build and verify
Change-Id: I7d1eb16a010dd6479514832e6e6333260b769bf8