leaudio: Fix chosing available configuration
When devices is the group are known, configuration are choosen for each
of the available context type based on the capabilities of the remote
devices.
Later during configuration, code is choosing best CIG paramerters
looking into remote capabilities and also by determining topology
(strategy).
This patch, adds strategy check when configuration is choosen for the
available context, so we are sure that the same configuration is
chosen later when CIG is created.

Bug: 248430016
Test: atest BluetoothInstrumentationTests
Test: atest bluetooth_le_audio_test
Test: manual stream to LeAudio device
Tag: #feature

Change-Id: Ifcb35b4554436a2061f03d95e757d5495bbb38c5