Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/956ca7be-a91a-4327-8679-40ac5493425c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c74e5524a43b9dd284e45d36109c7d9766cfc04