Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4f18acc-6b15-4571-bb5e-2ceb067d72df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iadc9d17859fd9bcb3b848b0dde1eed313eb44bc3