Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d82ec18-b8cd-47a4-bcdd-b09173579392). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd851788d13d0a4f5a7b2585348a8fea5935d48d