Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc541e9a-dba3-4501-bc38-857a21cc4dee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I879dbdb590d29373e070413b5e2f1f370a849ce0