Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81aea5f8-ecdc-4fb5-a231-d0a828fe84d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4939afa23df8d61867eb2d7703d160bd78a2dd26