Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/025a9616-b02b-4cea-ac3b-a7a8a32e83dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd9dcd1cf482f703e56c821a63c4d8b0bcb14823