Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cae6af1b-659f-4374-bbb0-120b551649ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I693e17e86c126975635f9d753f163dbd6e13c36b