Revert "floss: topshim: Deduplicate RawAddress definitions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d37befe5-c445-4644-917e-e38fcb25c381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b41382bf1f99611af2fbe9ae32ac1adb8f783d9