Revert "Create a gRPC client for HFP Service in Python to do top..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c981eb8e-4eaf-44e2-80c3-89ec8fad4ecf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I934c42376592fbc9cbd3ae0500e49d9649d8efd7