Revert "Create a gRPC client for HFP Service in Python to do top..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/129c0b6d-f34a-4139-999b-03700a6c62fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibdd1dd5c906070ed0600930e3864f3b7cd24940a