Revert "Send active device changed intent after connecting in Au..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a08728a4-fd46-48e0-bce3-3e206aa75de1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5d0907dc1f99c233e3a26681b1c31cf7053ac28