Revert "Add openReceivedFile tests in BluetoothOppUtilityTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67db8a2a-bd85-4414-b1d9-ca6b80336f8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf85c80f092363c8147cc8583ab72f5280695ae3