Revert "Add openReceivedFile tests in BluetoothOppUtilityTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/52a0068a-e80d-417d-8d90-f707000f3d5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6851cf6800df1d39ca41cd7766769b379123e27