Fix BR/EDR device may be flush by ble device
[Root Cause]
if there are too much ble device(adv), br/edr will be overwritten
by ble device, then targe device will not in inquiry db.
then it will not do RNR for some device without name.

[Solution]
so enlarge the inquiry db and reserve half of it for BR/EDR and ble.

Bug: 266219643
Test: atest net_test_btif_stack pass

Change-Id: I955fb52696ad9cdacc0aff7cbb4bf2dd81f342b0