Revert "vendor_libs: Check the HIDL status"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d758cc6a-a9ee-463f-8755-3ecc61e5c4fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a237b65135ebd664a2369ee89ba68cb9fbc03c1