Floss: pass request status for audio connection
Requests to start the audio are non-blocking and are not guaranteed to
be fulfilled. Currently, Floss exposes |get_x_audio_connected| to allow
the audio stack to query the audio state and decide if it should retry
or give up on the request in case it takes too long.

However, requests could be rejected immediately (e.g., due to Floss
being temporarily busy with other requests), in which case the audio
stack has no way to know of, and could only unavailingly query if the
audio state has changed to |started|.

In this CL, we return whether the request was made/being processed for
those interfaces.

Bug: 259420892
Test: Deploy and verify DBus signature

Change-Id: Ib9bb53f6b1a808bcfcaafd61fb550281cb33cb2e