Avoid out-of-order L2CAP CoC callbacks on channel creation
When L2CA_ConnectLECocReq is called and starts the channels state
machine, we do so synchronously. If the state machine experiences a
failure and invokes a callback, it does so again synchronously, so the
failure callback arrives *before* the connection request completes.

By deferring the state transition until after the connect request
returns, we fix the order of events.

Bug: 258223088
Test: TODO
Change-Id: I698c0e5e514485b9af8e21f136b5aa7c918a6d4c