Revert "Replace Log Statements in system/stack"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6841af44-4d70-4517-a461-d05c5b6539f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id130238d7cb53fe92538c4eaeed5c685c7ff9e09