Revert "Replace Log Statements in system/stack"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6841af44-4d70-4517-a461-d05c5b6539f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70dd3494c875f5586b87c29f0595a7172c6d5e9a