avrcp: uprev control service for target profile from 1.3 to 1.4
Many peripherals (A2DP source, a.k.a, AVRCP controller), when connected
to the Android AVRCP profile, rely on the Android Target implementation
and the Controller SDP record to progress correctly into the AVRCP
capabilities exchange phase. The existing code register an AVRCP
Controler service with version 1.3, which is not sufficient for BlueZ
based device as BlueZ requires to see an AVRCP profile with version 1.4
or higher to initiate the absolute volume interface.

Bug: 259144894
Tag: #floss
Test: bluetooth_AdapterAUHealth.au_avrcp_command_test.floss
Test: bluetooth_AdapterAUHealth.au_avrcp_media_info_test.floss
Test: gd/cert/run
Change-Id: Ie9597ca7bb5e0bc9f62c467eabbb446a5abb9a9c