LeAudio: Reduce number of reconfigurations
To play sonification events we don't really need high quality
stream and we can play them at any configuration. For example
a button click when resuming recording should not trigger
reconfiguration to Media.

This change allows to play these events without the reconfiguration
if these types of events are the only contributors to the current
audio stream (using allow list of contexts).

Bug: 261417889
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine  --no-bazel-mode
Change-Id: I5dbd66e3a4ef0944894550f9902cddf9d0422681