LeAudio: Use direction aware audio contexts on enable and metadata update
Passing sink and source audio contexts to the state machine as one
value is usually fine, but at the moment ASEs are enabled or metadata
is updated we should filter this combined value by the directional context
availability. For a mixed audio contexts this will prevent situations where
MEDIA context is used to enable remote audio source ASE (microphone).

Bug: 256967923
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test --no-bazel-mode
Change-Id: I675aa85ef5d3ed62703f5d5ce6e7d2d1083cf015