Update connection reporting to be based on peer role
Connection state reporting current checks which service is up without
any respect for the role the device is taking. This is because it
assumes that only one A2DP role can be active on the device at a time.
This needs to change in order to support concurrent A2DP roles.

Instead, we should check the role of the device and then sanity check
that the relevant A2DP role is active.

Tag: #stability
Bug: 259734525
Test: atest BluetoothInstrumentationTests
Change-Id: I4384a377e563968a03610d41d582a4cef6f156c8