Floss: propagate HFP audio disconnection with state validation
It has been observed that in some cases, the audio state callback for
HFP could be notified of disconnection before it was ever connected, and
then be notified of connection following after.

Within these two state callbacks, the audio stack may be confused and
ask for reconnection, causing state de-sync.

In this CL, we do not propagate the audio disconnection event to the
audio stack unless it is a meaningful state transition (i.e., connected
-> disconnected).

Bug: 260618573
Tag: #floss
Test: Build and verify

Change-Id: I033abde1bded53a8c7366a961daa6888a1e5b9b3