Floss: Avoid LE connection for classic devices
Open a GATT client creates an LE connection with peer, if this happens in
the pairing process with an classic keyboard with wrong passkey,
the bluetooth controller might not be reset or not be able to connect with
other peer devices. This patch avoid to connect/disconnect to battery
service for classic devices.

Bug: 258283982
Tag: #floss
Test: Manual - Pair CL keyboard with wrong key and then pair with LE mouse
Change-Id: I8d7e94058bda2a0f840f090f15d33b6a32fb2181