wip: pdl: Always apply read/write casts
This is necessary now that we have support for non-scalar types, such
as enums. Even if we read a chunk of exactly 16 bits, we might now
have to cast the integer to an enum.

There is a problem with this: we generate a lot of extra and
unnecessary code when dealing with scalars.

Test: atest pdl_tests pdl_rust_generator_tests_{le,be}
Change-Id: Iccaa3e4a545f91aa23afd59931e09e4944d3428b