Revert "Prevent null pointer exception in getCurrentCall method"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3229397-e7fb-443d-8156-327ea26c321d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie33a18d8bb9386d4387e8c6fae27bdc0f434515a