Floss: defer AVRCP connection as initiator
At the moment, collision happens very frequently because we initiate the
connection to AVRCP almost at the same time as connecting to the other
profiles, and most HFs initiate the same AVRCP connection as a response.

In this CL, we defer the AVRCP connection as initiator to the point
where it should only happen as a fallback when it is regarded as a
missing profile after the timeout, and A2DP is connected.

Bug: 260111031
Test: Deploy and verify all profiles connect consistently
Change-Id: I2a4637bed440ebf242087299595ae5633b9a0d7e