Revert "Add VolumeControlNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/28cda468-d02d-46ff-88b9-63c94e100031). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ide6dcbf6adeb1001abac95f538aa7addcfe9f864