Revert "Add VolumeControlNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8389c19-bbc0-4a9c-b1bd-db3aa3fde43d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b638aa035827d90ab7b8178e11ef43888986816