Revert "Add VolumeControlNativeInterfaceTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6548ac3-6f9a-4359-82f4-8ba1beae8c84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b309a9148d135f3a5d71c4874616c641cdb7981