Revert "Move TestHciLayer into a separate file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd7a517a-02b8-4210-bcf3-c534e019891f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida7d073453195d0392dd7a01bbe9e3bcb1bcdbb2