Revert^2 Move TestHciLayer into a separate file
Previously we were copy/pasting it everywhere. Not only was this huge code
duplication, it also violated ODR since we were not always correctly namespacing
it. This CL wraps all its copies inside an anonymous namespace, and also exposes
it in bluetooth::hci in a separate file.

Re-landing due to build failure on previous attempt, since presubmits passed but then a conflicting change was submitted later.

Change-Id: I42aa96a6415a509e85500cfab445f2497b2ca436
Test: all existing
Bug: 247839849