Fix unbroadcasted ACTION_ACL_CONNECTED Intent
The intent BluetoothDevice#ACTION_ACL_CONNECTED was never broadcasted for
a new incoming ACL connection. The device was normally added into the
remoteDevice map only if it was previously bonded which was preventing
the intent to being broadcasted in case the device is a new incoming
connection and stop at the ACL level.

Bug:
Test: avatar_runner --build example.py config.yml -- --tests=test_wait_connection
Change-Id: I80c5d7f26587fff2729351bb8a7c8f0e44524283