Revert "[Invisalign] Build changes to split out libbluetooth-core"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73f01154-7b62-4633-9cd0-e3dc3f410b8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bf7985dc8ebf570912ef11a2edd0a0ff77061dd