Revert "[Invisalign] Build changes to split out libbluetooth-core"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efcb67d8-d9eb-4853-832b-42b11fdc5a6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I51dc4f776969c662c0b12a32bf7b4cc46e20b120