Revert "snoop_logger: Add support for logging to a socket"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a635469-63e0-49e9-a722-25f7de92b163). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ca5023f573e70c4bfde07c1a26cac1959cd0fbb