Revert "Add BluetoothOppReceiverTest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bae3a8a3-de80-4e06-a923-6d403f5a2f81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76f1671649da19f27a011e0d7fb96e3c3a152bf7